+++ towerzap/towerzap.go
@@ -5,7 +5,6 @@ import (
 
 	"github.com/tigorlazuardi/tower"
 	"go.uber.org/zap"
-	"go.uber.org/zap/zapcore"
 )
 
 var _ tower.Logger = (*Logger)(nil)
@@ -37,8 +36,13 @@ func (l *Logger) SetTraceCapturer(capturer TraceCapturer) {
 }
 
 func (l Logger) Log(ctx context.Context, entry tower.Entry) {
-	elements := make([]zap.Field, 0, 7)
+	elements := make([]zap.Field, 0, 16)
+	elements = append(elements, zap.Time("time", entry.Time()))
 	elements = append(elements, l.tracer.CaptureTrace(ctx)...)
+	elements = append(elements, zap.Object("service", service(entry.Service())))
+	if key := entry.Key(); key != "" {
+		elements = append(elements, zap.String("key", key))
+	}
 	code := entry.Code()
 	if code != 0 {
 		elements = append(elements, zap.Int("code", code))
@@ -49,24 +53,7 @@ func (l Logger) Log(ctx context.Context, entry tower.Entry) {
 	if len(data) == 1 {
 		elements = append(elements, toField("context", data[0]))
 	} else if len(data) > 1 {
-		field := zapcore.ArrayMarshalerFunc(func(ae zapcore.ArrayEncoder) error {
-			for _, value := range data {
-				switch value := value.(type) {
-				case tower.Fields:
-					return ae.AppendObject(fields(value))
-				case zapcore.ObjectMarshaler:
-					return ae.AppendObject(value)
-				case zapcore.ArrayMarshaler:
-					return ae.AppendArray(value)
-				case map[string]any:
-					return ae.AppendObject(fields(value))
-				default:
-					return ae.AppendReflected(value)
-				}
-			}
-			return nil
-		})
-		elements = append(elements, zap.Array("context", field))
+		elements = append(elements, zap.Array("context", encodeContextArray(entry.Context())))
 	}
 
 	l.Logger.Log(translateLevel(entry.Level()), entry.Message(), elements...)
@@ -74,72 +61,21 @@ func (l Logger) Log(ctx context.Context, entry tower.Entry) {
 
 func (l Logger) LogError(ctx context.Context, err tower.Error) {
 	elements := make([]zap.Field, 0, 7)
-	elements = append(elements, zap.Error(err))
+	elements = append(elements, zap.Time("time", err.Time()))
 	elements = append(elements, l.tracer.CaptureTrace(ctx)...)
+	elements = append(elements, zap.Object("service", service(err.Service())))
 	elements = append(elements, zap.Int("code", err.Code()))
-	elements = append(elements, zap.Object("caller", zapcore.ObjectMarshalerFunc(func(oe zapcore.ObjectEncoder) error {
-		caller := err.Caller()
-		oe.AddString("origin", caller.ShortOrigin())
-		oe.AddString("location", caller.String())
-		return nil
-	})))
-
+	elements = append(elements, zap.Stringer("caller", err.Caller()))
+	if key := err.Key(); key != "" {
+		elements = append(elements, zap.String("key", key))
+	}
 	data := err.Context()
 	if len(data) == 1 {
 		elements = append(elements, toField("context", data[0]))
 	} else if len(data) > 1 {
-		field := zapcore.ArrayMarshalerFunc(func(ae zapcore.ArrayEncoder) error {
-			for _, value := range data {
-				switch value := value.(type) {
-				case tower.Fields:
-					return ae.AppendObject(fields(value))
-				case zapcore.ObjectMarshaler:
-					return ae.AppendObject(value)
-				case zapcore.ArrayMarshaler:
-					return ae.AppendArray(value)
-				case map[string]any:
-					return ae.AppendObject(fields(value))
-				default:
-					return ae.AppendReflected(value)
-				}
-			}
-			return nil
-		})
-		elements = append(elements, zap.Array("context", field))
+		elements = append(elements, zap.Array("context", encodeContextArray(err.Context())))
 	}
+	origin := err.Unwrap()
+	elements = append(elements, toField("error", origin))
 	l.Logger.Log(translateLevel(err.Level()), err.Message(), elements...)
-}
-
-func toField(key string, value any) zap.Field {
-	switch value := value.(type) {
-	case tower.Fields:
-		return zap.Object(key, fields(value))
-	case zapcore.ObjectMarshaler:
-		return zap.Object(key, value)
-	case zapcore.ArrayMarshaler:
-		return zap.Array(key, value)
-	case map[string]any:
-		return zap.Object(key, fields(value))
-	default:
-		return zap.Any(key, value)
-	}
-}
-
-func translateLevel(lvl tower.Level) zapcore.Level {
-	switch lvl {
-	case tower.DebugLevel:
-		return zapcore.DebugLevel
-	case tower.InfoLevel:
-		return zapcore.InfoLevel
-	case tower.WarnLevel:
-		return zapcore.WarnLevel
-	case tower.ErrorLevel:
-		return zapcore.ErrorLevel
-	case tower.FatalLevel:
-		return zapcore.FatalLevel
-	case tower.PanicLevel:
-		return zapcore.FatalLevel
-	default:
-		return zapcore.InvalidLevel
-	}
 }